+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.7] - 2026-01-11
+
+### Fixed
+
+- **TTL Type Handling**: `set_ttl()` and `__init__()` now accept `int` (seconds) in addition to `timedelta`, fixing `AttributeError: 'int' object has no attribute 'total_seconds'`
+- **Integration Tests**: Fixed flaky gallery count assertions to use dynamic checks instead of hardcoded values
+
 ## [0.10.6] - 2026-01-10
 
 ### Fixed
@@ -588,7 +595,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - respx for GraphQL HTTP mocking
 - 70%+ test coverage requirement
 
-[Unreleased]: https://github.com/Jakan-Kink/stash-graphql-client/compare/v0.6.0...HEAD
+[Unreleased]: https://github.com/Jakan-Kink/stash-graphql-client/compare/v0.10.7...HEAD
+[0.10.7]: https://github.com/Jakan-Kink/stash-graphql-client/compare/v0.10.6...v0.10.7
+[0.10.6]: https://github.com/Jakan-Kink/stash-graphql-client/compare/v0.10.5...v0.10.6
+[0.10.5]: https://github.com/Jakan-Kink/stash-graphql-client/compare/v0.10.4...v0.10.5
+[0.10.4]: https://github.com/Jakan-Kink/stash-graphql-client/compare/v0.10.3...v0.10.4
+[0.10.3]: https://github.com/Jakan-Kink/stash-graphql-client/compare/v0.10.2...v0.10.3
+[0.10.2]: https://github.com/Jakan-Kink/stash-graphql-client/compare/v0.10.1...v0.10.2
+[0.10.1]: https://github.com/Jakan-Kink/stash-graphql-client/compare/v0.10.0...v0.10.1
+[0.10.0]: https://github.com/Jakan-Kink/stash-graphql-client/compare/v0.9.0...v0.10.0
+[0.9.0]: https://github.com/Jakan-Kink/stash-graphql-client/compare/v0.8.2...v0.9.0
+[0.8.2]: https://github.com/Jakan-Kink/stash-graphql-client/compare/v0.8.1...v0.8.2
+[0.8.1]: https://github.com/Jakan-Kink/stash-graphql-client/compare/v0.8.0...v0.8.1
+[0.8.0]: https://github.com/Jakan-Kink/stash-graphql-client/compare/v0.7.2...v0.8.0
+[0.7.2]: https://github.com/Jakan-Kink/stash-graphql-client/compare/v0.7.1...v0.7.2
+[0.7.1]: https://github.com/Jakan-Kink/stash-graphql-client/compare/v0.7.0...v0.7.1
+[0.7.0]: https://github.com/Jakan-Kink/stash-graphql-client/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/Jakan-Kink/stash-graphql-client/compare/v0.5.1...v0.6.0
 [0.5.1]: https://github.com/Jakan-Kink/stash-graphql-client/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/Jakan-Kink/stash-graphql-client/compare/v0.5.0b7...v0.5.0